switch to new usage pattern.
an union is used to be able to statically alloc a struct big enough
to contain all 3 backend types rather than having only one type
baked into it. at this point all backends are interchangeable,
as long as the optional routines needed for SDL screen initialization
are used in the code using it (console_init_graphics).
it's a noop on the raw terminal backends.
a slight disadvantage of this approach is that a font need to be
compiled, even if the used terminal backend doesn't use it.
thus it'll probably be baked somehow directly into the sdl-backend
later.