ncconsole.c: fix mouse handling.
i previously assumed that has_mouse() would return true whenever
the terminal supports it and ncurses was built with appropriate
support code.

however, unlike has_colors(), this one only returns true when the
mousemode has been switched on using mousemask.

since you can see from the mousemask return value if it failed,
it seems the has_mouse() function is entirely useless.

additionally, mouse release events return the wrong bits set.
i left the old code intact, in case it is fixed in another ncurses
version, and added a workaround that reports a mouse_release
event like it should.

it seems it is impossible to get mouse movement events from the
terminals, without relying on gpm or similar.
thus you won't get any mouse move events currently.

as for termbox, we have to implement mouse handling ourselves in
a fork, as the author is not willing to add it:

https://github.com/nsf/termbox-go/pull/4

from this it seems we only have to add "\033[?1003h" to
activate the mouse mode on xterms, and "\033[?1003l" to turn it
off again. for linux console, s/1003/1000/.

the response then looks like "\033[M", followed by: b,x,y;
each of them an uint8_t.
b AND 3 returns the button number, and b AND 0x60 has some yet
to figure out special meaning; probably related to whether
the button was released or pressed.