switch to bitfont usage
this needs further cleanup.
to improve memory efficiency, we use only bitfonts now,
but not in the old dosfonts.c format which is bitswapped for some
reason.

gen/ now includes scripts and code to reproduce the new bitfont
format from the old sources. i.e. if one wants to edit the
testfont and update it. the old format makes editing a font
directly in the header convenient, the new one impossible.

since the bit shuffling is probably much slower than the direct
char access we had earlier, we buffer the last used font in a
static array and reuse that if the font hasn't changed.
it is expected that this saves 1/2 - 2/3 of conversions in typical
code.

everything seems to compile and work so far, so i commit it as-is,
and continue to cleanup later.