do all graphics/terminal init in console_init_graphics()
console_init() now only zeroes the struct and sets the backendtype.
this makes it possible to query the backend type before the terminal
is put into some weird mode.
we need this for programs that want to read from stdin, in case
the backend is a terminal, they have to dup the fd to prevent EOF
sent when the terminal wants to read later user input.