add console_settitle
some people may want to display something else than "sdl-console" in the title
of their app if the sdl backend is in use. also handy for debug info.
this is currently unimplemented in the ncurses and termbox backends,
but adding it there as well shouldn't be hard, since there's the escape seq.
"\x1b]2;TITLE\x07" which at least some terminals support. there's however no
official ncurses function to achieve it.