fix console_getbounds() in ncconsole
this was probably broken since 70a12c1e54a10ac03f97358a8632e573487f2243.
i assumed the values of _maxx, _maxy would be equivalent to what's returned
from getmaxyx() but that's apparently not the case. the documentation of
ncurses' getmaxyx() says:

  ...the WINDOW._maxx and WINDOW._maxy values in ncurses have (at least since
  release 1.8.1) differed by one from some other implementations.
  The difference is hidden by means of the macro getmaxyx

that is probably the reason we traditionally had it incremented by one.