sdlconsole: add a cache
the intention is to speed up operation at the cost of higher memory
usage. IIRC there was a slight performance increase when using it.
this code was written months ago, when i get to it i will make some
benchmarks and either make the code optional (with #ifdef), remove it,
or keep it.